refactor fea tag checks to a new custom fb profile
