+++ Makefile
@@ -93,6 +93,7 @@ test-fb-static-expert:
 	@echo " fontbakery v`fontbakery --version` static font checks"
 	@echo "========================================================="
 	fontbakery check-profile -C --loglevel WARN qa/check-googlesans.py $(EXPERT_STATIC_BUILD_DIR)/*.ttf
+	fontbakery check-profile -C --loglevel WARN qa/check-fea.py $(EXPERT_STATIC_BUILD_DIR)/*.ttf
 	fontbakery check-profile -C --loglevel WARN qa/check-charset.py $(EXPERT_STATIC_BUILD_DIR)/*.ttf
 
 test-fb-vf-expert:
@@ -101,6 +102,7 @@ test-fb-vf-expert:
 	@echo "========================================================="
 	# default build checks
 	fontbakery check-profile -C --loglevel WARN qa/check-googlesans.py $(EXPERT_VARIABLE_BUILD_DIR)/*.ttf
+	fontbakery check-profile -C --loglevel WARN qa/check-fea.py $(EXPERT_VARIABLE_BUILD_DIR)/*.ttf
 	fontbakery check-profile -C --loglevel WARN qa/check-charset.py $(EXPERT_VARIABLE_BUILD_DIR)/*.ttf
 	# non-default build checks
 	#  - partial instances do not repeat char set checks that were executed above in default build checks
