+++ qa/check-googlesans.py
@@ -31,10 +31,6 @@ GOOGLESANS_PROFILE_CHECKS = UNIVERSAL_PROFILE_CHECKS + [
     "com.google.fonts/check/googlesans/opentype/os2/typodescender",
     "com.google.fonts/check/googlesans/opentype/os2/typoascender",
     "com.google.fonts/check/googlesans/opentype/os2/typolinegap",
-    "com.google.fonts/check/googlesans/features/staticuprights",
-    "com.google.fonts/check/googlesans/features/staticitalics",
-    "com.google.fonts/check/googlesans/features/variableuprights",
-    "com.google.fonts/check/googlesans/features/variableitalics",
     "com.google.fonts/check/googlesans/vf/fvardefault",
 ]
 
@@ -65,137 +61,6 @@ ATTRIBUTES = {
     "wght_axis_default": 400.0,
 }
 
-STATIC_UPRIGHT_FEA = [
-    "aalt",
-    "c2sc",
-    "calt",
-    "case",
-    "ccmp",
-    "dlig",
-    "dnom",
-    "frac",
-    "kern",
-    "liga",
-    "lnum",
-    "locl",
-    "mark",
-    "mkmk",
-    "numr",
-    "ordn",
-    "pnum",
-    "sinf",
-    "smcp",
-    "ss01",
-    "ss02",
-    "ss03",
-    "ss04",
-    "ss05",
-    "ss06",
-    "ss07",
-    "ss08",
-    "subs",
-    "sups",
-    "tnum",
-]
-
-STATIC_ITALICS_FEA = [
-    "aalt",
-    "c2sc",
-    "calt",
-    "case",
-    "ccmp",
-    "dlig",
-    "dnom",
-    "frac",
-    "kern",
-    "liga",
-    "lnum",
-    "locl",
-    "mark",
-    "mkmk",
-    "numr",
-    "ordn",
-    "pnum",
-    "sinf",
-    "smcp",
-    "ss01",
-    "ss02",
-    "ss04",
-    "ss05",
-    "ss06",
-    "ss07",
-    "ss08",
-    "subs",
-    "sups",
-    "tnum",
-]
-
-VAR_UPRIGHT_FEA = [
-    "aalt",
-    "c2sc",
-    "calt",
-    "case",
-    "ccmp",
-    "dlig",
-    "dnom",
-    "frac",
-    "kern",
-    "liga",
-    "lnum",
-    "locl",
-    "mark",
-    "mkmk",
-    "numr",
-    "ordn",
-    "pnum",
-    "rvrn",
-    "sinf",
-    "smcp",
-    "ss01",
-    "ss02",
-    "ss03",
-    "ss04",
-    "ss05",
-    "ss06",
-    "ss07",
-    "ss08",
-    "subs",
-    "sups",
-    "tnum",
-]
-
-VAR_ITALICS_FEA = [
-    "aalt",
-    "c2sc",
-    "calt",
-    "case",
-    "ccmp",
-    "dlig",
-    "dnom",
-    "frac",
-    "kern",
-    "liga",
-    "lnum",
-    "locl",
-    "mark",
-    "mkmk",
-    "numr",
-    "ordn",
-    "pnum",
-    "rvrn",
-    "sinf",
-    "smcp",
-    "ss01",
-    "ss02",
-    "ss04",
-    "ss05",
-    "ss06",
-    "ss07",
-    "ss08",
-    "subs",
-    "sups",
-    "tnum",
-]
 
 # ================================================
 #
@@ -454,136 +319,6 @@ def com_google_fonts_check_googlesans_opentype_os2_typolinegap(ttFont):
 
 
 # ================================================
-# Feature support
-# ================================================
-
-# statics
-@check(
-    id="com.google.fonts/check/googlesans/features/staticuprights",
-    conditions=["is_not_italic", "is_not_variable_font"],
-    rationale="""
-    Confirms that the upright builds contain expected feature tags.
-    """,
-)
-def com_google_fonts_check_googlesans_features_static_uprights(ttFont):
-    """Confirms that the upright builds contain expected feature tags."""
-    tt = ttFont
-    gpos = tt["GPOS"]
-    gsub = tt["GSUB"]
-
-    fea_tags = set()
-
-    for gpos_record in gpos.table.FeatureList.FeatureRecord:
-        fea_tags.add(gpos_record.FeatureTag)
-
-    for gsub_record in gsub.table.FeatureList.FeatureRecord:
-        fea_tags.add(gsub_record.FeatureTag)
-
-    if sorted(fea_tags) == STATIC_UPRIGHT_FEA:
-        yield PASS, f"{tt.reader.file.name} contains the expected feature tags"
-    else:
-        yield (
-            FAIL,
-            f"{tt.reader.file.name} does not contain the expected feature tags.\n"
-            f"Found:{sorted(fea_tags)}\nExpected:{STATIC_UPRIGHT_FEA}",
-        )
-
-
-@check(
-    id="com.google.fonts/check/googlesans/features/staticitalics",
-    conditions=["is_italic", "is_not_variable_font"],
-    rationale="""
-    Confirms that the italics builds contain expected feature tags.
-    """,
-)
-def com_google_fonts_check_googlesans_features_static_italics(ttFont):
-    """Confirms that the italics builds contain expected feature tags."""
-    tt = ttFont
-    gpos = tt["GPOS"]
-    gsub = tt["GSUB"]
-
-    fea_tags = set()
-
-    for gpos_record in gpos.table.FeatureList.FeatureRecord:
-        fea_tags.add(gpos_record.FeatureTag)
-
-    for gsub_record in gsub.table.FeatureList.FeatureRecord:
-        fea_tags.add(gsub_record.FeatureTag)
-
-    if sorted(fea_tags) == STATIC_ITALICS_FEA:
-        yield PASS, f"{tt.reader.file.name} contains the expected feature tags"
-    else:
-        yield (
-            FAIL,
-            f"{tt.reader.file.name} does not contain the expected feature tags.\n"
-            f"Found:{sorted(fea_tags)}\nExpected:{STATIC_ITALICS_FEA}",
-        )
-
-
-# VF
-@check(
-    id="com.google.fonts/check/googlesans/features/variableuprights",
-    conditions=["is_not_italic", "is_variable_font"],
-    rationale="""
-    Confirms that the variable upright builds contain expected feature tags.
-    """,
-)
-def com_google_fonts_check_googlesans_features_variable_uprights(ttFont):
-    """Confirms that the upright builds contain expected feature tags."""
-    tt = ttFont
-    gpos = tt["GPOS"]
-    gsub = tt["GSUB"]
-
-    fea_tags = set()
-
-    for gpos_record in gpos.table.FeatureList.FeatureRecord:
-        fea_tags.add(gpos_record.FeatureTag)
-
-    for gsub_record in gsub.table.FeatureList.FeatureRecord:
-        fea_tags.add(gsub_record.FeatureTag)
-
-    if sorted(fea_tags) == VAR_UPRIGHT_FEA:
-        yield PASS, f"{tt.reader.file.name} contains the expected feature tags"
-    else:
-        yield (
-            FAIL,
-            f"{tt.reader.file.name} does not contain the expected feature tags.\n"
-            f"Found:{sorted(fea_tags)}\nExpected:{VAR_UPRIGHT_FEA}",
-        )
-
-
-@check(
-    id="com.google.fonts/check/googlesans/features/variableitalics",
-    conditions=["is_italic", "is_variable_font"],
-    rationale="""
-    Confirms that the variable italics builds contain expected feature tags.
-    """,
-)
-def com_google_fonts_check_googlesans_features_variable_italics(ttFont):
-    """Confirms that the italics builds contain expected feature tags."""
-    tt = ttFont
-    gpos = tt["GPOS"]
-    gsub = tt["GSUB"]
-
-    fea_tags = set()
-
-    for gpos_record in gpos.table.FeatureList.FeatureRecord:
-        fea_tags.add(gpos_record.FeatureTag)
-
-    for gsub_record in gsub.table.FeatureList.FeatureRecord:
-        fea_tags.add(gsub_record.FeatureTag)
-
-    if sorted(fea_tags) == VAR_ITALICS_FEA:
-        yield PASS, f"{tt.reader.file.name} contains the expected feature tags"
-    else:
-        yield (
-            FAIL,
-            f"{tt.reader.file.name} does not contain the expected feature tags.\n"
-            f"Found:{sorted(fea_tags)}\nExpected:{VAR_ITALICS_FEA}",
-        )
-
-
-# ================================================
 # Variable build format specific
 # ================================================
 
@@ -614,8 +349,7 @@ def com_google_fonts_check_googlesans_variable_fvar_default(ttFont):
             else:
                 yield (
                     PASS,
-                    f"{tt.reader.file.name} contains the expected fvar "
-                    f"opsz default.",
+                    f"{tt.reader.file.name} contains the expected fvar " f"opsz default.",
                 )
         elif axis.axisTag == "wght":
             if axis.defaultValue != EXPECTED_WGHT:
@@ -628,8 +362,7 @@ def com_google_fonts_check_googlesans_variable_fvar_default(ttFont):
             else:
                 yield (
                     PASS,
-                    f"{tt.reader.file.name} contains the expected fvar "
-                    f"opsz default.",
+                    f"{tt.reader.file.name} contains the expected fvar " f"opsz default.",
                 )
